Create Registering compliant with LG Customer care - SAMPLE v1
Greet the customer and explain the purpose of the call.
Ask the customer to provide the model number and serial number of their LG washing machine.
Verify the authenticity of the customer by asking them to provide the last four digits of their credit card number or other personal information that was used during the purchase of the washing machine.
Once the customer's identity is verified, ask them to describe the malfunction in detail.
Based on the description of the malfunction, determine the appropriate category for the complaint (e.g. mechanical issue, electrical issue, etc.).
Create a new case in the LG CRM system and assign it to the appropriate team or individual.
Generate a unique case ID for the complaint.
Read out the case ID to the customer and confirm that they have received it.
Thank the customer for contacting LG and assure them that their complaint will be addressed as soon as possible.
Follow up with the customer as necessary to resolve the issue and close the case in the CRM system.